// src/balances/dtos/balance.dto.ts
export class BalanceDTO {
	id?: string;
	created_at?: Date;
	updated_at?: Date;
	last_balance_id?: string;
	gold_weight?: number;
	cash_amount?: number;
	bill_million_count: number;
	bill_100k_count: number;
	bill_50k_count: number;
	bill_10k_count: number;
	bill_5k_count: number;
	bill_1k_count: number;
	bill_500_count: number;
	bill_100_count: number;
	bill_50_count: number;
	bill_10_count: number;
	retail_999_weight?: number;
	retail_9999_weight?: number;
	factory_995_weight?: number;
	factory_999_weight?: number;
	factory_9999_weight?: number;
	old_pure_gold_weight?: number;
	old_mixed_gold_weight?: number;
	product_gold_weight?: number;
	advanced_gold_weight?: number;
	advanced_cash_amount?: number;
	unpaid_gold_weight?: number;
	unpaid_cash_amount?: number;
	lend_gold_weight: number;
	borrow_gold_weight: number;
}